Move Gio action helpers to own file
